+++ src/utils.js
@@ -9,28 +9,16 @@ export function isObject(value) {
 }
 
 /**
- * Adds or subtracts date portions to the given date and returns the new date.
+ * Validates if the key is not empty.
+ * (null, undefined or empty string)
  *
- * @see https://gist.github.com/jherax/bbc43e479a492cc9cbfc7ccc20c53cd2
- *
- * @param  {object} options: It contains the date parts to add or remove, and can have the following properties:
- *         - {Date} date: if provided, this date will be affected, otherwise the current date will be used.
- *         - {number} minutes: minutes to add/subtract
- *         - {number} hours: hours to add/subtract
- *         - {number} days: days to add/subtract
- *         - {number} months: months to add/subtract
- *         - {number} years: years to add/subtract
- * @return {Date}
+ * @param  {string} key: keyname of an element in the storage mechanism
+ * @return {void}
  */
-export function alterDate(options = {}) {
-  const opt = Object.assign({}, options);
-  const d = opt.date instanceof Date ? opt.date : new Date();
-  if (+opt.minutes) d.setMinutes(d.getMinutes() + opt.minutes);
-  if (+opt.hours) d.setHours(d.getHours() + opt.hours);
-  if (+opt.days) d.setDate(d.getDate() + opt.days);
-  if (+opt.months) d.setMonth(d.getMonth() + opt.months);
-  if (+opt.years) d.setFullYear(d.getFullYear() + opt.years);
-  return d;
+export function checkEmpty(key) {
+  if (key == null || key === '') {
+    throw new Error('The key provided can not be empty');
+  }
 }
 
 /**
@@ -54,14 +42,17 @@ export function setProperty(obj, name, value) {
 }
 
 /**
- * Validates if the key is not empty.
- * (null, undefined or empty string)
+ * Try to parse a value from JSON.
  *
- * @param  {string} key: keyname of an element in the storage mechanism
- * @return {void}
+ * @param  {string} value: the value to parse
+ * @return {any}
  */
-export function checkEmpty(key) {
-  if (key == null || key === '') {
-    throw new Error('The key provided can not be empty');
+export function tryParse(value) {
+  let parsed;
+  try {
+    parsed = JSON.parse(value);
+  } catch (e) {
+    parsed = value;
   }
+  return parsed;
 }